feat(go): add api to trigger license validation; move license functions and types into separate package;
